added rake task to make OSApp ready OSW/OSM combo
The goal was to use OS App as OSW editor/runner for situations where I don't want users to have to use the command line and CLI. I had hoped I did not have to have measures copied into workflow in run directory but I couldn't get it to run with remote measure location. I need to update the rest of the template OSW files to have name field for viewing in OS App.